chore: get average elapsed time per host
